refactor(service): export internal packages for downstream test isolation
Move fixtures and subjectmappingbuiltin out of service/internal/ so they
can be imported from a separate integration-test module (DSPX-2582).
This eliminates the testcontainers-go transitive dependency leak into
downstream consumers by enabling integration tests to live in their own
Go module.

- service/internal/fixtures/ → service/pkg/testutil/fixtures/
- service/internal/subjectmappingbuiltin/ → service/pkg/subjectmappingbuiltin/
- Update all import paths and YAML load paths
