+++ service/integration/keymanagement_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/opentdf/platform/protocol/go/policy/kasregistry"
 	"github.com/opentdf/platform/protocol/go/policy/keymanagement"
 	"github.com/opentdf/platform/protocol/go/policy/unsafe"
-	"github.com/opentdf/platform/service/internal/fixtures"
 	"github.com/opentdf/platform/service/pkg/db"
+	"github.com/opentdf/platform/service/pkg/testutil/fixtures"
 	"github.com/stretchr/testify/suite"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
