+++ service/integration/namespaces_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/opentdf/platform/protocol/go/policy/kasregistry"
 	"github.com/opentdf/platform/protocol/go/policy/namespaces"
 	"github.com/opentdf/platform/protocol/go/policy/unsafe"
-	"github.com/opentdf/platform/service/internal/fixtures"
 	"github.com/opentdf/platform/service/pkg/db"
+	"github.com/opentdf/platform/service/pkg/testutil/fixtures"
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/protobuf/proto"
 )
