+++ service/integration/obligations_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/opentdf/platform/protocol/go/common"
 	"github.com/opentdf/platform/protocol/go/policy"
 	"github.com/opentdf/platform/protocol/go/policy/obligations"
-	"github.com/opentdf/platform/service/internal/fixtures"
 	"github.com/opentdf/platform/service/pkg/db"
+	"github.com/opentdf/platform/service/pkg/testutil/fixtures"
 	"github.com/stretchr/testify/suite"
 	"google.golang.org/protobuf/proto"
 )
